Optimize safety for State.enterContext and State.exitContext.  Also, optimize performance for State.setContextValue.  Add more comments and clean up/rename unit test.
