+++ Src/java/engine/src/main/java/org/opencds/cqf/cql/engine/elm/executing/ExpressionDefEvaluator.java
@@ -8,8 +8,9 @@ import org.opencds.cqf.cql.engine.execution.State;
 
 public class ExpressionDefEvaluator {
     public  static Object internalEvaluate(ExpressionDef expressionDef, State state, ElmLibraryVisitor<Object,State> visitor) {
+        boolean isEnteredContext = false;
         if (expressionDef.getContext() != null) {
-            state.enterContext(expressionDef.getContext());
+            isEnteredContext = state.enterContext(expressionDef.getContext());
         }
         try {
             state.pushEvaluatedResourceStack();
@@ -20,7 +21,6 @@ public class ExpressionDefEvaluator {
                 return er.value();
             }
 
-            // LUKETODO:  we have the expression name here
             Object value = visitor.visitExpression(expressionDef.getExpression(), state);
 
             if (state.getCache().isExpressionCachingEnabled()) {
@@ -32,9 +32,8 @@ public class ExpressionDefEvaluator {
         }
         finally {
             state.popEvaluatedResourceStack();
-            if (expressionDef.getContext() != null) {
-                state.exitContext();
-            }
+            // state.enterContext.getContext() == null will result in isEnteredContext = false, which means pop() won't be called
+            state.exitContext(isEnteredContext);
         }
     }
 }
